Fix: Only print "Could not stop reading pipeline bus" if the connection had a timeout.
The `is_reading_bus` queue can already hung up, which is not an error.
This can happen when the connection is closed really fast and the sender from the `is_reading_bus` was already freed.
This causes the `is_reading_bus.recv_timeout(BUS_EOS_TIMEOUT)` to fail, because the connection hung up.
But that doesn't mean that the pipeline couldn't stop successfully.
Printing an error is misleading in that case. Only printing an error if the timeout has reached, solves that our tests are failing.

Fixes #48