Add tests pipeline to the ci
Changes:
- Add tests to the ci
- Reduce the sleep duration in the tests to 2 seconds, so the tests are running faster
- Use logging warn instead of error for the max amount of shown participants

The talk is using the loglevel error when more users are trying to be shown in the recorder, which exceed the limit of maximum participants for the recorder.
This loglevel causes the tests to fail, because the tests are failing when an error log was produced.
Using the loglevel warn makes more sense in that case and would fix the tests.