Resolve "Running the compositor tests is causing some gstreamer error"
Especially the `test_remove()` had a problem with re-occurring stream IDs.
The test creates eight streams and removes them.
It does this 50 times.

So after a test round has ended with all eight streams (ID `0`-`7`) removed, the test then created new streams with the same IDs again.

This was irritating gstreamer somehow.

And several further changes like:

- remove the unnecessary EOS check when dropping the mixer
- update some tests to set a speaker
- review `Stream` a bit
- add queues behind tees
- some debug improvements

Also tried to use blocking probes but that doesn't work proper.
Must use timeouts for waiting for blocking probes which is reached in significant amount in tests.
Also it's needed to block two pads for audio and video which is not described how to do in gstreamer's dynamic pipeline doc.
Maybe this is something to ask Sebastian.
So decided to remove the probes until this seems helpful again.

Closes #70