docs: fix some typos and remove license headers
Some tools such as docusaurus are unable to properly render the page
titles in the page index when the title is not in the first line of the
file. License for the affected files can be declared e.g. in a `.reuse/
dep5` file using the `docs/**.md` pattern in the future.