feature: Prioritize screen capture over camera feed
The speaker detection is showing the current speaker user at the top.
This can be annoying by sharing their screen.
Now the screen capture has a higher priority, if someone has a screen capture running
and another person without a screen capture is talking, then the main view will still be the screen capture.
If a second person is sharing their screen and is talking, then the second users screen capture
will replace the main view.
If no one is sharing their screen, the video feed of the speaking use will be the main view.

Closes #33