feat: Move Multisink into compositor
It's mandatory to remove the Multisink and move the `tee` into the `Mixer`, to be able to dynamically start and stop a sink.
The upcoming streaming capability needs to dynamically add and removed Sink outputs (for example Start/Stop a streaming service). It was impossible with the old solution to add dynamically `Sinks`. This is a problem with object safety traits. A solution we had was to move the `tee` part inside the compositor and handle the adding/removing of sinks in there, instead of the `Multisink`.

The `Mixer` startup was also split into an `AudioMixer` and `VideoMixer`, which is now easier to handle and easier to change.