fix: remove obsolete usage of `media` signaling module
This code just extracted data from the `media` namespace if present in
some messages, but didn't do anything beyond collecting the data in a
variable which would then go out of scope. Cleaning this leftover
removes the `opentalk-types` dependency which was of no good use any
more and is already deleted in the OpenTalk controller repository.